+++ packages/secrets/core/src/os/mac/keychain_search.rs
@@ -1,7 +1,8 @@
 use crate::os::mac::error::{handle_os_status, Error};
 use crate::os::mac::ffi::{
-    kSecAttrAccount, kSecAttrLabel, kSecAttrService, kSecClass, kSecClassGenericPassword,
-    kSecMatchLimit, kSecReturnAttributes, kSecReturnData, kSecReturnRef, SecItemCopyMatching,
+    kSecAttrAccount, kSecAttrLabel, kSecAttrService, kSecClass, kSecClassCertificate,
+    kSecClassGenericPassword, kSecClassIdentity, kSecMatchLimit, kSecReturnAttributes,
+    kSecReturnData, kSecReturnRef, SecItemCopyMatching, kSecMatchSubjectContains,
 };
 use crate::os::mac::keychain_item::SecKeychainItem;
 use crate::os::mac::misc::{SecCertificate, SecIdentity, SecKey};
@@ -22,6 +23,8 @@ pub struct KeychainSearch {
     label: Option<CFString>,
     service: Option<CFString>,
     account: Option<CFString>,
+    subject_contains: Option<CFString>,
+    class: Option<CFTypeRef>,
     load_attrs: bool,
     load_data: bool,
     load_refs: bool,
@@ -145,6 +148,25 @@ impl KeychainSearch {
         self
     }
 
+    pub fn subject_contains(&mut self, subject: &str) -> &mut Self {
+        self.subject_contains = Some(CFString::new(subject));
+        self
+    }
+
+    pub fn class_identity(&mut self) -> &mut Self {
+        unsafe {
+            self.class = Some(kSecClassIdentity.cast());
+        }
+        self
+    }
+
+    pub fn class_certificate(&mut self) -> &mut Self {
+        unsafe {
+            self.class = Some(kSecClassCertificate.cast());
+        }
+        self
+    }
+
     pub fn with_attrs(&mut self) -> &mut Self {
         self.load_attrs = true;
         self
@@ -169,16 +191,26 @@ impl KeychainSearch {
         let mut params = vec![];
 
         unsafe {
-            params.push((
-                CFString::wrap_under_get_rule(kSecClass),
-                CFType::wrap_under_get_rule(kSecClassGenericPassword.cast()),
-            ));
+            // Use configured class if set, otherwise default to generic password
+            let class_value = if let Some(class) = self.class {
+                CFType::wrap_under_get_rule(class)
+            } else {
+                CFType::wrap_under_get_rule(kSecClassGenericPassword.cast())
+            };
 
-            // Handle any parameters that were configured before execution (label, service, account)
+            params.push((CFString::wrap_under_get_rule(kSecClass), class_value));
+
+            // Handle any parameters that were configured before execution (label, service, account, subject)
             if let Some(ref label) = self.label {
                 params.push((
                     CFString::wrap_under_get_rule(kSecAttrLabel),
                     label.as_CFType(),
+                ));
+            }
+            if let Some(ref subject) = self.subject_contains {
+                params.push((
+                    CFString::wrap_under_get_rule(kSecMatchSubjectContains),
+                    subject.as_CFType(),
                 ));
             }
             if let Some(ref service) = self.service {