+++ packages/secrets/src/keyring/index.js
@@ -57,6 +57,7 @@ const {
     findPassword,
     getPassword,
     setPassword,
+    createTlsPipe,
 } = requireFn(binaryPath);
 
 module.exports.deletePassword = deletePassword;
@@ -64,3 +65,4 @@ module.exports.findCredentials = findCredentials;
 module.exports.findPassword = findPassword;
 module.exports.getPassword = getPassword;
 module.exports.setPassword = setPassword;
+module.exports.createTlsPipe = createTlsPipe;