+++ __tests__/__packages__/cli-test-utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe CLI test utils package will be documented in this file.
 
+## `8.32.0`
+
+- BugFix: Replaced `uuid` dependency for technical currency. [#2720](https://github.com/zowe/zowe-cli/pull/2720)
+
 ## `8.7.1`
 
 - BugFix: Improved the error message shown on MacOS when `runCliScript` method fails to run script that is missing shebang line. [#2314](https://github.com/zowe/zowe-cli/pull/2314)